feat: add prependFolderNameToID
finds the folder name of the file and prepends that to the ID.
e.g. an ID inside the component folder `Footer` with an id of: `.button` will be
converted to `Footer.button`. The feature have been made opt-in with the use of
`.`. This enables the feature to gradually added, and allows developers to opt
out in case of a folder rename.

Make sure `extractSourceLocation` is set to `true` in your `babel-plugin-react-intl`
config when using this feature